chore: organize test files, fix key issue
